Proyecto Version Final

+++ bd1-todo-main/bd1-todo-app/bd1-todo-core/src/main/java/tec/bd/todo/Todo.java
@@ -5,7 +5,6 @@ import tec.bd.todo.repository.TodoRepository;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
-import java.util.UUID;
 
 public class Todo {
 
@@ -25,6 +24,9 @@ public class Todo {
 
     public TodoRecord getById(String id) {
         // TODO: validar que el id no sea nulo y si es nulo lanzar una excepcion
+        if(id==null){
+            throw new RuntimeException("IllegalArgumentException");
+        }
         return this.todoRepository.findById(id);
     }
 
@@ -34,48 +36,46 @@ public class Todo {
         if(null == record.getStartDate()) {
             record.setStartDate(new Date());
         }
-        record.setId(UUID.randomUUID().toString());
         record.setStatus(Status.NEW);
-        return this.todoRepository.save(record);
-    }
-
-    public TodoRecord update(TodoRecord record) {
-        Objects.requireNonNull(record);
-
-        // TODO: si el todoRecord.Id no existe, lanzar exception
-        // el registro a actualizar tiene que existir.
-
-        // Si el titulo es nulo, devolver exception
-        if(null == record.getStartDate()) {
-            record.setStartDate(new Date());
+        if (this.todoRepository.findById(record.getId()) != null &  record.getTitle()!= null) {
+            return this.todoRepository.save(record);
         }
-
-        return this.todoRepository.update(record);
+        else{
+            throw new RuntimeException("IllegalArgumentException");
+        }
     }
 
     public void deleteTodoRecord(String id) {
-
-        // TODO: buscar si el record existe, y si existe borrarlo
-        this.todoRepository.remove(id);
+        if (this.todoRepository.findById(id) != null) {
+            this.todoRepository.remove(id);
+        }
+        else{
+            throw new RuntimeException("IllegalArgumentException");
+        }
     }
 
-    public List<TodoRecord> getStartDateRange(Date startDate, Date endDate) {
-        // TODO: validamos que startDAte y endDate no sean nulos
-        // TODO: validamos que el endDate es mayor o igual que el startDate
-        // TODO: validamos que las fechas tienen que estar en un rango definido por alguien
-        return this.todoRepository.findByBetweenStartDates(startDate, endDate);
+    public List<TodoRecord> getStartDateRange(Date startDate, Date endDate){
+        if(startDate==null || endDate==null){
+            throw new RuntimeException("IllegalArgumentException");
+        }
+      return this.todoRepository.findByBetweenStartDates(startDate,endDate);
     }
-
-    public List<TodoRecord> searchInTitle(String textToSearch) {
-        // TODO: validar que el textToSearch no sea nulo ni vacio
+    public List<TodoRecord> searchInTitle(String textToSearch){
+        if(textToSearch==null){
+            throw new RuntimeException("IllegalArgumentException");
+        }
         return this.todoRepository.findByPatternInTitle(textToSearch);
     }
 
-    public TodoRecord updateTodoRecord(TodoRecord todoRecord) {
-        // TODO: que el ID del todoRecord Exista
-        // TODO: validar que como minimo el todoRecord tenga ID, titulo, startDate y Status
-        // TODO: el titulo, startDate, Status no pueden nulos (o vacios)
-        return this.todoRepository.update(todoRecord);
+
+    public TodoRecord updateTodoRecord(TodoRecord todoRecord){
+        if (this.todoRepository.findById(todoRecord.getId()) != null && todoRecord.getTitle()!= null && todoRecord.getStartDate()!=null && todoRecord.getStatus()!=null) {
+            return this.todoRepository.update(todoRecord);
+        }
+        throw new RuntimeException("IllegalArgumentException");
     }
+
+
+
 
 }
